Make it build on AIX
git-svn-id: http://svn.code.sf.net/p/xmlrpc-c/code@322 98333e67-4a24-44d7-a75c-e53540dd3050

+++ trunk/xmlrpc-c/src/xmlrpc_builddecomp.c
@@ -694,7 +694,7 @@ readDatetimeStr(xmlrpc_env *         const envP,
 static void
 readBase64(xmlrpc_env *           const envP,
            const xmlrpc_value *   const valueP,
-           unsigned int *         const lengthP,
+           size_t *               const lengthP,
            const unsigned char ** const byteStringValueP,
            xmlrpc_bool            const oldstyleMemMgmt) {
