+++ trunk/xmlrpc-c/src/xmlrpc_int.h
@@ -135,7 +135,7 @@ xmlrpc_read_string_w_lp_old(xmlrpc_env *     const envP,
 void
 xmlrpc_read_base64_old(xmlrpc_env *           const envP,
                        const xmlrpc_value *   const valueP,
-                       unsigned int *         const lengthP,
+                       size_t *               const lengthP,
                        const unsigned char ** const byteStringValueP);
